Ntrlab

+++ backend/app/controllers/users.js
@@ -66,7 +66,7 @@ module.exports = {
                 }).returning(Token.body));
                 bologger.log({
                     //req: req, Does not use req if you want to use public namespace TODO realm?
-                    user: req.user.id,
+                    user: req.user,
                     action: 'insert',
                     object: 'token',
                     entities: {
@@ -138,7 +138,7 @@ module.exports = {
         }).then(function(){
             bologger.log({
                 //req: req, Does not use req if you want to use public namespace TODO realm?
-                user: req.user.id,
+                user: req.user,
                 action: 'delete',
                 object: 'token',
                 entities: {
@@ -234,6 +234,36 @@ module.exports = {
 
             var user = yield thunkQuery(User.insert(newClient).returning(User.id));
 
+            bologger.log({
+                //req: req, Does not use req if you want to use public namespace TODO realm?
+                user: req.user,
+                action: 'insert',
+                object: 'users',
+                entity: _.first(user).id,
+                info: 'Add new superuser (invite)'
+            });
+
+            var essenceId = yield * common.getEssenceId(req, 'Users');
+            var notifyLevel = 2; // ToDo: Default - need specify notifyLevel in frontend
+            var note = yield * notifications.createNotification(req,
+                {
+                    userFrom: req.user.realmUserId ? req.user.realmUserId : userId,
+                    userTo: _.first(user).id,
+                    body: 'Superadmin Invite',
+                    essenceId: essenceId,
+                    entityId: _.first(user).id,
+                    notifyLevel: notifyLevel,
+                    name: req.body.firstName,
+                    surname: req.body.lastName,
+                    login: req.body.email,
+                    password: pass,
+                    token: activationToken,
+                    subject: 'Indaba. Superadmin invite',
+                    config: config
+                },
+                'invite'
+            );
+
             return user;
 
         }).then(function (data) {
@@ -280,7 +310,7 @@ module.exports = {
     //            userId = yield thunkQuery(User.insert(newClient).returning(User.id));
     //            bologger.log({
     //                //req: req, Does not use req if you want to use public namespace TODO realm?
-    //                user: req.user.id,
+    //                user: req.user,
     //                action: 'insert',
     //                object: 'users',
     //                entity: _.first(userId).id,
@@ -296,7 +326,7 @@ module.exports = {
     //            var organizationId = yield thunkQuery(Organization.insert(newOrganization).returning(Organization.id));
     //            bologger.log({
     //                //req: req, Does not use req if you want to use public namespace TODO realm?
-    //                user: req.user.id,
+    //                user: req.user,
     //                action: 'insert',
     //                object: 'organizations',
     //                entity: _.first(organizationId).id,
@@ -314,7 +344,7 @@ module.exports = {
     //            );
     //            bologger.log({
     //                //req: req, Does not use req if you want to use public namespace TODO realm?
-    //                user: req.user.id,
+    //                user: req.user,
     //                action: 'insert',
     //                object: 'projects',
     //                entity: _.first(projectId).id,
@@ -328,7 +358,7 @@ module.exports = {
     //            }));
     //            bologger.log({
     //                //req: req, Does not use req if you want to use public namespace TODO realm?
-    //                user: req.user.id,
+    //                user: req.user,
     //                action: 'update',
     //                object: 'users',
     //                entity: _.first(userId).id,
@@ -402,7 +432,7 @@ module.exports = {
             var updated = yield thunkQuery(User.update(data).where(User.activationToken.equals(req.params.token)).returning(User.id));
             bologger.log({
                 req: req,
-                user: _.first(updated).id,
+                user: _.first(updated),
                 action: 'update',
                 object: 'users',
                 entity: _.first(updated).id,
@@ -463,7 +493,7 @@ module.exports = {
             }
             bologger.log({
                 req: req,
-                user: req.user.realmUserId,
+                user: req.user,
                 action: 'update',
                 object: 'organizations',
                 entity: _.first(updated).id,
@@ -493,9 +523,11 @@ module.exports = {
             if (!req.body.email || !req.body.firstName) {
                 throw new HttpError(400, 'Email and First name fields are required');
             }
+
             if (!vl.isEmail(req.body.email)) {
                 throw new HttpError(400, 101);
             }
+
             var isExistUser = yield thunkQuery(User.select(User.star()).where(User.email.equals(req.body.email)));
             isExistUser = _.first(isExistUser);
 
@@ -534,15 +566,26 @@ module.exports = {
                 };
 
                 var userId = yield thunkQuery(User.insert(newClient).returning(User.id));
+
                 newUserId = userId[0].id;
                 bologger.log({
                     req: req,
-                    user: req.user.realmUserId,
+                    user: req.user,
                     action: 'insert',
                     object: 'users',
                     entity: newUserId,
                     info: 'Add new user (org invite)'
                 });
+
+                if (req.body.roleID == 2) { // invite admin
+                    if (!org.adminUserId) {
+                        yield thunkQuery(
+                            Organization.update({adminUserId:newUserId}).where(Organization.id.equals(org.id))
+                        );
+                    }
+                }
+
+
             }else {
                 newClient = isExistUser;
             }
@@ -610,7 +653,7 @@ module.exports = {
         }).then(function(data){
             bologger.log({
                 req: req,
-                user: req.user.realmUserId,
+                user: req.user,
                 action: 'insert',
                 object: 'UserUOA',
                 entities: {
@@ -640,7 +683,7 @@ module.exports = {
         }).then(function(data){
             bologger.log({
                 req: req,
-                user: req.user.realmUserId,
+                user: req.user,
                 action: 'delete',
                 object: 'UserUOA',
                 entities: {
@@ -674,7 +717,7 @@ module.exports = {
         }).then(function (data) {
             bologger.log({
                 req: req,
-                user: req.user.realmUserId,
+                user: req.user,
                 action: 'delete',
                 object: 'UserUOA',
                 entities: data,
@@ -725,7 +768,7 @@ module.exports = {
         }).then(function (data) {
             bologger.log({
                 req: req,
-                user: req.user.realmUserId,
+                user: req.user,
                 action: 'insert',
                 object: 'UserUOA',
                 entities: data,
@@ -770,16 +813,18 @@ module.exports = {
         var thunkQuery = req.thunkQuery;
         co(function*(){
             var updateObj = _.pick(req.body, User.whereCol);
+            console.log(updateObj.password);
             if(updateObj.password){
                 updateObj.password = User.hashPassword(updateObj.password);
             }
+            console.log(updateObj.password);
             if (Object.keys(updateObj).length) {
                 yield thunkQuery(
                     User.update(updateObj).where(User.id.equals(req.params.id))
                 );
                 bologger.log({
                     req: req,
-                    user: req.user.realmUserId,
+                    user: req.user,
                     action: 'update',
                     object: 'users',
                     entity: req.params.id,
@@ -791,7 +836,7 @@ module.exports = {
             );
             bologger.log({
                 req: req,
-                user: req.user.realmUserId,
+                user: req.user,
                 action: 'delete',
                 object: 'userGroups',
                 entities: userGroups4delete,
@@ -813,7 +858,7 @@ module.exports = {
                 );
                 bologger.log({
                     req: req,
-                    user: req.user.realmUserId,
+                    user: req.user,
                     action: 'insert',
                     object: 'userGroups',
                     entities: groupObjs,
@@ -833,13 +878,22 @@ module.exports = {
         var thunkQuery = req.thunkQuery;
 
         co(function* (){
+            var adminUser = yield thunkQuery(Organization.
+                select(Organization.adminUserId)
+                .from(Organization)
+                .where(Organization.realm.equals(req.params.realm))
+            );
+            if (_.first(adminUser) && _.first(adminUser).adminUserId === parseInt(req.params.id)) {
+                throw new HttpError(400, 'It does not possible delete organization`s admin');
+            }
+
             return yield thunkQuery(
                 User.delete().where(User.id.equals(req.params.id))
             );
         }).then(function(data){
             bologger.log({
                 req: req,
-                user: req.user.realmUserId,
+                user: req.user,
                 action: 'delete',
                 object: 'users',
                 entity: req.params.id,
@@ -936,7 +990,7 @@ module.exports = {
         }).then(function(){
             bologger.log({
                 req: req,
-                user: req.user.realmUserId,
+                user: req.user,
                 action: 'update',
                 object: 'users',
                 entity: req.user.id,
@@ -1036,7 +1090,7 @@ module.exports = {
 
             bologger.log({
                 req: req,
-                user: user.realmUserId,
+                user: user,
                 action: 'update',
                 object: 'users',
                 entity: user.id,
@@ -1120,7 +1174,7 @@ module.exports = {
         }).then(function (data) {
             bologger.log({
                 //req: req, Does not use req if you want to use public namespace TODO realm?
-                user: _.first(data).id,
+                user: _.first(data),
                 action: 'update',
                 object: 'users',
                 entity: _.first(data).id,
@@ -1250,7 +1304,7 @@ function* insertOne(req, res, next) {
     var user = yield thunkQuery(User.insert(req.body).returning(User.id));
     bologger.log({
         req: req,
-        user: req.user.realmUserId,
+        user: req.user,
         action: 'insert',
         object: 'users',
         entity: _.first(user).id,
